+++ cmd/daemon/main.go
@@ -30,7 +30,7 @@ import (
 )
 
 var (
-	version   = "v2.0.1"
+	version   = "v2.1.0"
 	commit    = "none"
 	buildDate = "unknown"
 )